+++ test/ganglion/timingEventsAsPromises.js
